+++ src/Interop/NativeImageFormat.cs
@@ -15,16 +15,10 @@ namespace GmicSharp.Interop
     {
         // 8-bit grayscale with no alpha channel
         Gray8 = 0,
-        // Opaque 24-bit color using the BGR format, 8 bits per component
-        Bgr888,
-        // Opaque 32-bit color using the BGR format, 8 bits per component
-        Bgr888x,
-        // 32-bit color using the BGRA format, 8 bits per component
-        Bgra8888,
+        // 8-bit grayscale with an 8-bit alpha channel
+        GrayAlpha88,
         // Opaque 24-bit color using the RGB format, 8 bits per component
         Rgb888,
-        // Opaque 32-bit color using the RGB format, 8 bits per component
-        Rgb888x,
         // 32-bit color using the RGBA format, 8 bits per component
         Rgba8888
     }